Delete common
