+++ src/elements/text-input/src/index.tsx
@@ -1,42 +1,77 @@
-import * as React from 'react'
+/** @jsx jsx */
 
-import { Input, Width } from './common'
+import { useState } from 'react'
+import { jsx } from '@emotion/core'
+import { FrozenInput } from '@uswitch/trustyle.frozen-input'
+import { inputs } from '@uswitch/trustyle.styles'
+import InputMask from 'react-input-mask'
 
-export interface CommonInputProps {
+import * as st from './styles'
+
+export type InputType = 'text' | 'email' | 'tel' | 'date'
+export type Width = 'half' | 'full'
+
+export interface Props extends React.InputHTMLAttributes<HTMLInputElement> {
   freezable?: boolean
   hasError?: boolean
-  label?: string
-}
-export interface TextInputProps extends React.InputHTMLAttributes<HTMLInputElement>, CommonInputProps {
   mask?: string
-  maxLength?: number
   name: string
-  value?: string
+  value: string | undefined
   width?: Width
 }
 
-export interface DateInputProps extends CommonInputProps {
-  maxDate?: string
-  minDate?: string
-  name: string
+export const Input: React.FC<Props> = ({
+  freezable,
+  hasError = false,
+  mask,
+  width = 'full',
+  ...inputProps
+}) => {
+  const stringValue = inputProps.value || ''
+  const [hasFocus, setHasFocus] = useState(false)
+  const onBlur = (event: React.FocusEvent<HTMLInputElement>) => {
+    setHasFocus(false)
+
+    if (inputProps.onBlur) {
+      inputProps.onBlur(event)
+    }
+  }
+  const onFocus = (event: React.FocusEvent<HTMLInputElement>) => {
+    setHasFocus(true)
+
+    if (inputProps.onFocus) {
+      inputProps.onFocus(event)
+    }
+  }
+
+  const childProps = {
+    ...inputProps,
+    css: inputs.keyboardInput,
+    onBlur,
+    onFocus
+  }
+
+  return (
+    <FrozenInput text={stringValue} freezable={freezable}>
+      <div css={[inputs.keyboardInputContainer(hasError, hasFocus), st[width]]}>
+        {mask ? (
+          <InputMask mask={mask} {...childProps} value={stringValue} />
+        ) : (
+          <input {...childProps} value={stringValue} />
+        )}
+      </div>
+    </FrozenInput>
+  )
 }
 
-export interface EmailInputProps extends CommonInputProps {
-  name: string
-}
-
-export const TextInput: React.FC<TextInputProps> = props => (
+export const TextInput: React.FC<Props> = props => (
   <Input {...props} type="text" />
 )
 
-export const DateInput: React.FC<DateInputProps> = props => (
-  <Input {...props} type="date" />
-)
-
-export const EmailInput: React.FC<EmailInputProps> = props => (
+export const EmailInput: React.FC<Props> = props => (
   <Input {...props} type="email" />
 )
 
-export const TelInput: React.FC<TextInputProps> = props => (
+export const TelInput: React.FC<Props> = props => (
   <Input {...props} type="tel" />
 )
