+++ src/elements/text-input/src/stories.tsx
@@ -6,8 +6,6 @@ import { TextInput, TelInput } from './.'
 storiesOf('Elements|Text Input', module).add('example', () => (
   <div>
     <TextInput
-      onBlur={() => {}}
-      onChange={() => {}}
       name="example"
       value="Cascat"
     />
@@ -20,27 +18,17 @@ storiesOf('Elements|Text Input', module).add('example', () => (
     />
     <TextInput
       hasError
-      onBlur={() => {}}
-      onChange={() => {}}
       name="example"
       value="Error!"
     />
     <TelInput
       mask="99-99-99"
       name="example"
-      onBlur={() => {}}
-      onChange={() => {}}
       value="Masked input"
     />
     <TelInput
       name="example"
-      onBlur={() => {}}
-      onChange={() => {}}
-      dataProps={{
-        pii: true,
-        text: 'some-text',
-        nullable: null
-      }}
+      data-pii
       value="Data props"
     />
   </div>
